ddr_umctl: Fix warning due to recent type change of 'info.size'

+++ src/ddr_umctl.c
@@ -596,7 +596,7 @@ int ddr_init(const struct ddr_config *cfg)
 
 	VERBOSE("name = %s\n", cfg->info.name);
 	VERBOSE("speed = %d kHz\n", cfg->info.speed);
-	VERBOSE("size  = %zdM\n", cfg->info.size / 1024 / 1024);
+	VERBOSE("size  = %zdM\n", (size_t) cfg->info.size / 1024 / 1024);
 
 	/* Reset, start clocks at desired speed */
 	ddr_reset(cfg, true);
